resource_help/requirements.txt
appdirs==1.4.3
boto3==1.4.4
botocore==1.5.57
docutils==0.13.1
jmespath==0.9.3
olefile==0.44
packaging==16.8
Pillow==4.1.1
pyparsing==2.2.0
python-dateutil==2.6.0
s3transfer==0.1.10
six==1.10.0
